+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for Office 365 workbooks
 - Added support for OpenDocument spreadsheets in ods and fods format
 
 ## [2.1.0] - 2019-03-25